backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
elasticsearch==8.11.1
python-dotenv==1.0.1
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0 